Level 3 is done
I went with Dice's Coefficient for finding string similarity. This
algorithm returns a similarity between 0 and 1. The more similar,
the higher the number. I decided on a threshold of .8 because it
filtered out different people with the same domain part of the email.
This treshold can easily be adjusted in the component.